pkgsite style docs
