+++ statemachine/cb/callbacks.go
@@ -1,0 +1,134 @@
+package cb
+
+import (
+	"ergo.services/actor/statemachine/action"
+	"ergo.services/ergo/gen"
+)
+
+/*
+Type alias for handlers called upon messages generated by [gen.Process.Send] (MessageHandlers).
+
+Each message handler receives as its inputs:
+  - sender's [gen.PID]
+  - current statemachine state
+  - current data associated to it
+  - message of type M that was received
+  - and a pointer to statemachine process
+
+Following Erlang's phylosopy, callback handler returns a number of values:
+  - a state to which statemachine should be transitioned (newState). It may be the same as current state.
+  - an updated Data object
+  - a list of [Action]'s to perform before switching to newState
+  - an error object
+
+If err is a non-nil object, the actor terminates. You can use [gen.TerminateReasonNormal]
+as usual to terminate the actor gracefully.
+*/
+type StateMessageHandler[D any, M any] func(
+	sender gen.PID,
+	state gen.Atom,
+	data D,
+	message M,
+	p gen.Process,
+) (
+	newState gen.Atom,
+	newData D,
+	actions []action.Action,
+	err error,
+)
+
+/*
+Type alias for handlers called upon messages generated by [gen.Process.Call] (CallHandlers).
+
+Each message handler receives as its inputs:
+  - sender's [gen.PID]
+  - current statemachine state
+  - current data associated to it
+  - message of type M that was received
+  - and a pointer to statemachine process
+
+Following Erlang's phylosopy, callback handler returns a number of values:
+  - a state to which statemachine should be transitioned (newState). It may be the same as current state.
+  - an updated Data object
+  - a list of [Action]'s to perform before switching to newState
+  - result, that will be send back to the caller in the synchronous manner
+  - an error object
+
+If err is a non-nil object, the actor terminates. You can use [gen.TerminateReasonNormal]
+as usual to terminate the actor gracefully.
+*/
+type StateCallHandler[D any, M any, R any] func(
+	sender gen.PID,
+	state gen.Atom,
+	data D,
+	message M,
+	p gen.Process,
+) (
+	newState gen.Atom,
+	newData D,
+	result R,
+	actions []action.Action,
+	err error,
+)
+
+// Type alias for event handler callbacks.
+// D is the type of the data associated with the StateMachine.
+// E is the type of the event.
+
+/*
+Type alias for handlers called upon events generated by [gen.Process.SendEvent] (EventHandlers).
+
+Each message handler receives as its inputs:
+  - current statemachine state
+  - current data associated to it
+  - message of type M that was received
+  - and a pointer to statemachine process
+
+Following Erlang's phylosopy, callback handler returns a number of values:
+  - a state to which statemachine should be transitioned (newState). It may be the same as current state.
+  - an updated Data object
+  - a list of [Action]'s to perform before switching to newState
+  - an error object
+
+If err is a non-nil object, the actor terminates. You can use [gen.TerminateReasonNormal]
+as usual to terminate the actor gracefully.
+*/
+type EventHandler[D any, E any] func(
+	state gen.Atom,
+	data D,
+	event E,
+	p gen.Process,
+) (
+	newState gen.Atom,
+	newData D,
+	actions []action.Action,
+	err error,
+)
+
+// Type alias for StateEnter callback.
+// D is the type of the data associated with the StateMachine.
+
+/*
+Following the [gen_statem]'s phylosophy, there is a handler function, that is called upon
+every state change. The [state_enter/0].
+
+StateEnter handler receives as its inputs:
+  - current statemachine state
+  - target statemachine state
+  - current data associated to it
+  - and a pointer to statemachine process
+
+Transition may be overriden by returning newState different from toState.
+
+[state_enter/0]: https://www.erlang.org/doc/apps/stdlib/gen_statem.html#t:state_enter/0
+*/
+type StateEnterCallback[D any] func(
+	fromState gen.Atom,
+	toState gen.Atom,
+	data D,
+	p gen.Process,
+) (
+	newState gen.Atom,
+	newData D,
+	err error,
+)
