+++ statemachine/statemachine_test.go
@@ -1,8 +1,10 @@
-package statemachine
+package statemachine_test
 
 import (
 	"testing"
 
+	"ergo.services/actor/statemachine"
+	"ergo.services/actor/statemachine/action"
 	"ergo.services/ergo/gen"
 	"ergo.services/ergo/testing/unit"
 )
@@ -10,8 +12,13 @@ import (
 type Data struct{}
 
 type BasicStatemachine struct {
-	StateMachine[Data]
+	statemachine.StateMachine[Data]
 }
+
+const (
+	stateA = gen.Atom("StateA")
+	stateB = gen.Atom("StateB")
+)
 
 type StateChange struct{}
 
@@ -19,27 +26,27 @@ func factoryBasicStatemachine() gen.ProcessBehavior {
 	return &BasicStatemachine{}
 }
 
-func (b *BasicStatemachine) Init(args ...any) (StateMachineSpec[Data], error) {
-	spec := NewStateMachineSpec(
-		gen.Atom("StateA"),
+func (b *BasicStatemachine) Init(args ...any) (statemachine.StateMachineSpec[Data], error) {
+	spec := statemachine.NewStateMachineSpec(
+		stateA,
 
-		WithStateEnterCallback(stateEnter),
-		WithStateMessageHandler(gen.Atom("StateA"), changeState),
-		WithStateCallHandler(gen.Atom("StateA"), changeStateSync),
+		statemachine.WithStateEnterCallback(stateEnter),
+		statemachine.WithStateMessageHandler(stateA, changeState),
+		statemachine.WithStateCallHandler(stateA, changeStateSync),
 	)
 	return spec, nil
 }
 
-func changeState(state gen.Atom, data Data, msg StateChange, proc gen.Process) (gen.Atom, Data, []Action, error) {
-	return gen.Atom("StateB"), data, nil, nil
+func changeState(_ gen.PID, state gen.Atom, data Data, msg StateChange, proc gen.Process) (gen.Atom, Data, []action.Action, error) {
+	return stateB, data, nil, nil
 }
 
-func changeStateSync(state gen.Atom, data Data, msg StateChange, proc gen.Process) (gen.Atom, Data, gen.Atom, []Action, error) {
-	return gen.Atom("StateB"), data, gen.Atom("StateB"), nil, nil
+func changeStateSync(_ gen.PID, state gen.Atom, data Data, msg StateChange, proc gen.Process) (gen.Atom, Data, gen.Atom, []action.Action, error) {
+	return stateB, data, stateB, nil, nil
 }
 
 func stateEnter(oldState gen.Atom, newState gen.Atom, data Data, proc gen.Process) (gen.Atom, Data, error) {
-	if newState == gen.Atom("StateB") {
+	if newState == stateB {
 		return newState, data, gen.TerminateReasonNormal
 	}
 	return newState, data, nil
